+++ src/main/java/com/deepexi/support/amqp/listener/decorated/InvocableHandlerMethodDecoration.java
@@ -1,4 +1,4 @@
-package com.deepexi.support.amqp.listener;
+package com.deepexi.support.amqp.listener.decorated;
 
 import org.springframework.messaging.Message;
 import org.springframework.messaging.handler.invocation.InvocableHandlerMethod;
