+++ src/test/java/com/deepexi/support/amqp/integration/SimpleIntegrationTest.java
@@ -5,10 +5,10 @@ import com.deepexi.support.amqp.event.EventMessageTypeMapping;
 import com.deepexi.support.amqp.integration.event.BarEvent;
 import com.deepexi.support.amqp.integration.event.FooEvent;
 import com.deepexi.support.amqp.integration.event.RepeatedEvent;
-import com.deepexi.support.amqp.listener.Authenticator;
-import com.deepexi.support.amqp.listener.IdempotentValidator;
-import com.deepexi.support.amqp.listener.MessageRecorder;
-import com.deepexi.support.amqp.listener.SimpleMemoryIdempotentValidator;
+import com.deepexi.support.amqp.listener.decorator.Authenticator;
+import com.deepexi.support.amqp.listener.decorator.IdempotentValidator;
+import com.deepexi.support.amqp.listener.decorator.MessageRecorder;
+import com.deepexi.support.amqp.listener.decorator.SimpleMemoryIdempotentValidator;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.Ignore;
