+++ src/test/java/com/deepexi/support/amqp/listener/DecoratedMessageHandlerMethodFactoryTest.java
@@ -1,5 +1,8 @@
 package com.deepexi.support.amqp.listener;
 
+import com.deepexi.support.amqp.listener.decorated.DecoratedMessageHandlerMethodFactory;
+import com.deepexi.support.amqp.listener.decorated.InvocableHandlerMethodDecoration;
+import com.deepexi.support.amqp.listener.decorated.InvocableHandlerMethodDecorator;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.Before;
 import org.junit.Test;
